+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [6.0.0] - 2024-06-04
+
 ### Added
 
 - `googleMapsIntegrationKeyId` to `FormsAppBase`